Add patch extraction script and incorporate patches in dataloader and training, need to fix predictions
Patch extraction was finalised and added to a script. The dataloader and training were adapted to
incorporate the patches, but the predictions are not working with the full map. The model is now
expecting input sizes of (..., ..., patch_size, patch_size, ...) whereas for predictions we want
to give the full Swiss map.